requirements.txt
asn1crypto==0.24.0
astroid==2.0.4
cffi==1.11.5
cryptography==2.3.1
Django==2.1.1
django-filter==2.0.0
djangorestframework==3.8.2
djangorestframework-simplejwt==3.2.3
idna==2.7
isort==4.3.4
lazy-object-proxy==1.3.1
mccabe==0.6.1
mysqlclient==1.3.13
pycparser==2.18
PyJWT==1.6.4
pylint==2.1.1
PyMySQL==0.9.2
pytz==2018.5
PyYAML==3.13
six==1.11.0
wrapt==1.10.11
